Refactored instructions with masks for imm field and split tests into multiple files for different encoding types.

+++ include/rviss/isa/addi.h
@@ -6,7 +6,7 @@
 
 class ADDI : public InstructionTypeI {
 public:
-    ADDI(Register rd, Register rs, uint32_t imm);
+    ADDI(Register rd, Register rs, uint16_t imm);
 };
 
 #endif
