+++ include/rviss/isa/csrrsi.h
@@ -6,7 +6,7 @@
 
 class CSRRSI : public InstructionTypeI {
 public:
-    CSRRSI(Register rd, Register rs, uint32_t imm);
+    CSRRSI(Register rd, Register rs, uint16_t imm);
 };
 
 #endif
