+++ include/rviss/isa/slli.h
@@ -6,7 +6,7 @@
 
 class SLLI : public InstructionTypeI {
 public:
-    SLLI(Register rd, Register rs, uint32_t imm);
+    SLLI(Register rd, Register rs, uint16_t imm);
 };
 
 #endif
