+++ src/isa/addi.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/addi.h"
 
-ADDI::ADDI(Register rd, Register rs1, uint32_t imm)
+ADDI::ADDI(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::ADDI, rd, rs1, imm) {}
