+++ src/isa/andi.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/andi.h"
 
-ANDI::ANDI(Register rd, Register rs1, uint32_t imm)
+ANDI::ANDI(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::ANDI, rd, rs1, imm) {}
