+++ src/isa/fence.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/fence.h"
 
-FENCE::FENCE(uint32_t imm)
+FENCE::FENCE(uint16_t imm)
     : InstructionTypeI(Opcode::FENCE, Register::zero, Register::zero, imm) {}
