+++ src/isa/lh.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/lh.h"
 
-LH::LH(Register rd, Register rs1, uint32_t imm)
+LH::LH(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::LH, rd, rs1, imm) {}
