+++ src/isa/lhu.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/lhu.h"
 
-LHU::LHU(Register rd, Register rs1, uint32_t imm)
+LHU::LHU(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::LHU, rd, rs1, imm) {}
