+++ src/isa/lw.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/lw.h"
 
-LW::LW(Register rd, Register rs1, uint32_t imm)
+LW::LW(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::LW, rd, rs1, imm) {}
