+++ src/isa/sraiw.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/sraiw.h"
 
-SRAIW::SRAIW(Register rd, Register rs1, uint32_t imm)
+SRAIW::SRAIW(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::SRAIW, rd, rs1, imm) {}
